+++ CHANGELOG.md
@@ -6,7 +6,15 @@ All notable changes to this project will be documented in this file. See [Conven
 
 #### Features
 
-- [PR #133](https://github.com/wildpow/new-esc-gatsby/pull/133) landed!!!
+- [PR #135](https://github.com/wildpow/new-esc-gatsby/pull/135)
+  - Sheets now have a color picker instead of a variants being in one drop down.
+  - Redesign on single product form:
+    - Quantity input can no longer be less than 1 or greater than the maxQty prop.
+    - Qty and Submit display on the same line when only one variant exists.
+    - Single product form handling is now done by validation and disabling of form elements.
+    - Removal of dedicated error components.
+    - Added aria-labels to qty and sheet picker fields.
+- [PR #133](https://github.com/wildpow/new-esc-gatsby/pull/133)
   - Complete over hull of back-end. All product data is now in DatoCMS leaving Shopify to just deal with pricing.
   - Product features now have descriptions.
   - Foundations/frames/boxsprings have been added to accessories.